+++ lib/danger/request_sources/bitbucket_server.rb
@@ -49,8 +49,12 @@ module Danger
       def setup_danger_branches
         base_branch = self.pr_json[:toRef][:id].sub("refs/heads/", "")
         base_commit = self.pr_json[:toRef][:latestCommit]
+        # Support for older versions of Bitbucket Server
+        base_commit = self.pr_json[:toRef][:latestChangeset] if self.pr_json[:fromRef].key? :latestChangeset
         head_branch = self.pr_json[:fromRef][:id].sub("refs/heads/", "")
         head_commit = self.pr_json[:fromRef][:latestCommit]
+        # Support for older versions of Bitbucket Server
+        head_commit = self.pr_json[:fromRef][:latestChangeset] if self.pr_json[:fromRef].key? :latestChangeset
 
         # Next, we want to ensure that we have a version of the current branch at a known location
         scm.ensure_commitish_exists_on_branch! base_branch, base_commit